% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{add_questionnaire}
\alias{add_questionnaire}
\title{Add questionnaire}
\usage{
add_questionnaire(
  access_token,
  api_info,
  respondent_track_id,
  survey_id,
  round_description,
  valid_from,
  valid_untill,
  round_order = NULL,
  check_ssl = TRUE
)
}
\arguments{
\item{access_token}{API access token.}

\item{api_info}{List with api information.}

\item{respondent_track_id}{Respondent track identifier.}

\item{survey_id}{Survey identifier.}

\item{round_description}{Gems description of the round within a track.}

\item{valid_from}{Date from which the survey token should be valid.}

\item{valid_untill}{Date until which the survey token should be valid.}

\item{round_order}{Integer value to indicate the order of the round.}

\item{check_ssl}{Whether to check the SSL certificate or allow insecure
connections.}
}
\description{
Creates a survey task for a respondent in a specific track.
}
