% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{change_execution_period}
\alias{change_execution_period}
\title{Change execution period}
\usage{
change_execution_period(
  access_token,
  token_id,
  api_info,
  valid_from,
  valid_untill,
  check_ssl = TRUE
)
}
\arguments{
\item{access_token}{API access token.}

\item{token_id}{Identifier of the token to change.}

\item{api_info}{List with api information.}

\item{valid_from}{New date from which the survey token should be valid.}

\item{valid_untill}{New date until which the survey token should be valid.}

\item{check_ssl}{Whether to check the SSL certificate or allow insecure
connections.}
}
\description{
Change the validity of a Gems token.
}
