% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{decode_url_key}
\alias{decode_url_key}
\title{Decode url key}
\usage{
decode_url_key(hash_key, url_key, debug = FALSE)
}
\arguments{
\item{hash_key}{Key that is given to decrypt}

\item{url_key}{Encrypted key retrieved from the url}
}
\value{
String containing the encrypted password which can be used to get an
  access token
}
\description{
Decode the key in the url to get a password for the access
  token.
}
