% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{decompress_key}
\alias{decompress_key}
\title{decompress_key}
\usage{
decompress_key(key)
}
\arguments{
\item{key}{GZ compressed key.}
}
\value{
Decompressed key.
}
\description{
Decompress a gz compressed key.
}
