% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_access_token}
\alias{get_access_token}
\title{Get access token}
\usage{
get_access_token(api_info, check_ssl = TRUE, debug = FALSE)
}
\arguments{
\item{api_info}{List with api information.}

\item{check_ssl}{Whether to check the SSL certificate or allow insecure
connections.}

\item{debug}{Whether to enable debugging messages.}
}
\value{
Token to access the API.
}
\description{
Get access token
}
