% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_api_info}
\alias{get_api_info}
\title{Get api info}
\usage{
get_api_info(environment)
}
\arguments{
\item{environment}{One of: ["production", "acceptance", "testing"].}
}
\value{
List with api information.
}
\description{
Get api info
}
