% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_model_mapper}
\alias{get_model_mapper}
\title{Get model mapper}
\usage{
get_model_mapper(
  patient_number,
  respondent_track_id,
  organisation_id,
  access_token,
  base_model_mapper_url,
  dataset = "arat",
  check_ssl = TRUE,
  debug = FALSE
)
}
\arguments{
\item{patient_number}{Patient number, without the organisation.}

\item{respondent_track_id}{Respondent track identifier.}

\item{organisation_id}{Organisation identifier.}

\item{access_token}{API access token.}

\item{base_model_mapper_url}{API endpoint for model mapper data.}

\item{dataset}{Data set to retrieve data from.}

\item{check_ssl}{Whether to check the SSL certificate or allow insecure
connections.}

\item{debug}{Whether to enable debugging messages.}
}
\value{
Data.table containing model mapper data.
}
\description{
Get model mapper
}
