% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_new_token}
\alias{get_new_token}
\title{Get new token}
\usage{
get_new_token(access_token, old_token, api_info, check_ssl = TRUE)
}
\arguments{
\item{access_token}{API access token.}

\item{old_token}{Expired API access token.}

\item{api_info}{List with api information.}

\item{check_ssl}{Whether to check the SSL certificate or allow insecure
connections.}

\item{debug}{Whether to enable debugging messages.}
}
\value{
New API access token.
}
\description{
Get new token
}
