% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_organisations}
\alias{get_organisations}
\title{Get organisation (for given patient)}
\usage{
get_organisations(
  patient_number,
  organisation_id,
  access_token,
  base_organisation_url,
  check_ssl = TRUE,
  debug = FALSE
)
}
\arguments{
\item{patient_number}{Patient number, without the organisation.}

\item{organisation_id}{Organisation identifier.}

\item{access_token}{API access token.}

\item{base_organisation_url}{API endpoint for organisation data.}

\item{check_ssl}{Whether to check the SSL certificate or allow insecure
connections.}

\item{debug}{Whether to enable debugging messages.}
}
\value{
Data.table containing organisation data.
}
\description{
Get organisation (for given patient)
}
