% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_patient_info}
\alias{get_patient_info}
\title{Get patient info}
\usage{
get_patient_info(
  patient_id,
  access_token,
  base_patient_url,
  language = "nl",
  check_ssl = TRUE,
  debug = FALSE
)
}
\arguments{
\item{patient_id}{String formatted like "patient_id@organisation_id".}

\item{base_patient_url}{API endpoint for patient data.}

\item{language}{Language of the response data.}

\item{check_ssl}{Whether to check the SSL certificate or allow insecure
connections.}

\item{debug}{Whether to enable debugging messages.}
}
\value{
Data.table containing patient data.
}
\description{
Get patient info
}
