% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_query_data}
\alias{get_query_data}
\title{Get query data}
\usage{
get_query_data(
  url,
  token,
  output_type = "parsed",
  language = "nl",
  check_ssl = TRUE,
  debug = FALSE
)
}
\arguments{
\item{url}{The url to make your request to.}

\item{token}{API access token.}

\item{output_type}{One of ["parsed", "text"].}

\item{language}{Language of the response data.}

\item{check_ssl}{Whether to check the SSL certificate or allow insecure
connections.}

\item{debug}{Whether to enable debugging messages.}
}
\value{
List including the response data.
}
\description{
Get query data
}
