% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_responses}
\alias{get_responses}
\title{Get responses}
\usage{
get_responses(
  task_ids,
  access_token,
  base_response_url,
  language = "nl",
  check_ssl = TRUE,
  debug = FALSE
)
}
\arguments{
\item{task_ids}{Vector of task id's (same as token id's) to retrieve the
responses from.}

\item{access_token}{API access token.}

\item{base_response_url}{API endpoint for response data.}

\item{language}{Language of the response data.}

\item{check_ssl}{Whether to check the SSL certificate or allow insecure
connections.}

\item{debug}{Whether to enable debugging messages.}
}
\value{
Data.table containing response data.
}
\description{
Get responses
}
