% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R, R/utils.R
\name{get_return_url}
\alias{get_return_url}
\title{Get return url}
\usage{
get_return_url(api_info, patient_id, organisation_id, environment, short_url)

get_return_url(api_info, patient_id, organisation_id, environment, short_url)
}
\arguments{
\item{api_info}{List with api information.}

\item{patient_id}{Patient identifier.}

\item{organisation_id}{Organisation identifier.}

\item{environment}{One of: ["production", "acceptance", "testing"].}

\item{short_url}{Use short url format (for Pulse Next).}
}
\value{
String containing the encoded return url.

String containing the encoded return url.
}
\description{
Create a redirect link to return from Gems to the dashboard.

Create a redirect link to return from Gems to the dashboard.
}
