% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_task_info}
\alias{get_task_info}
\title{Get task info}
\usage{
get_task_info(
  patient_id,
  access_token,
  base_task_url,
  language = "nl",
  check_ssl = TRUE,
  debug = FALSE,
  tracks = NULL,
  careplan_ids = NULL
)
}
\arguments{
\item{patient_id}{String formatted like "patient_id@organisation_id".}

\item{access_token}{API access token.}

\item{base_task_url}{API endpoint for task data.}

\item{language}{Language of the response data.}

\item{check_ssl}{Whether to check the SSL certificate or allow insecure
connections.}

\item{debug}{Whether to enable debugging messages.}

\item{tracks}{Tracks to filter on.}

\item{careplan_ids}{Careplan id's to filter on.}
}
\value{
Data.table containing task info data.
}
\description{
Get task info
}
