% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{post_activity_log}
\alias{post_activity_log}
\title{Post activity log}
\usage{
post_activity_log(
  activity_log_url,
  body,
  access_token,
  language = "nl",
  check_ssl = TRUE,
  debug = FALSE
)
}
\arguments{
\item{activity_log_url}{API endpoint for the activity log.}

\item{body}{Content to post as a list.}

\item{access_token}{API access token.}

\item{language}{Language of the response data.}

\item{check_ssl}{Whether to check the SSL certificate or allow insecure
connections.}

\item{debug}{Whether to enable debugging messages.}
}
\description{
Post activity log
}
