% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{decode_url_key}
\alias{decode_url_key}
\title{decode_url_key}
\usage{
decode_url_key(hash_key, url_key, gunzip = FALSE, debug = FALSE)
}
\arguments{
\item{hash_key}{Key that is given to decrypt}

\item{url_key}{Encrypted key retrieved from the url}

\item{gunzip}{Whether to gunzip the key after decoding.}

\item{debug}{Whether to enable debugging messages.}
}
\value{
String containing the encrypted password which can be used to get an
  access token
}
\description{
Decode the key in the url to get a password for the access
  token.
}
